# dashboard/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "larkstack console API",
    "description": "Admin API for the larkstack host: subsystem status, live config, an event stream (SSE), the Lark-app registry, action dispatch, and Lark OAuth login.",
    "contact": {
      "name": "S3anJia"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/apps": {
      "get": {
        "tags": [
          "console"
        ],
        "summary": "`GET /api/apps` — registered App manifests, for generic UI rendering.",
        "operationId": "list_apps",
        "responses": {
          "200": {
            "description": "Registered app manifests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/config/{app}/lark-app": {
      "put": {
        "tags": [
          "console"
        ],
        "summary": "`PUT /api/config/{app}/lark-app` — bind (or clear) the `[lark-apps.<name>]`\nan app uses for Lark credentials, writing just `[<app>].lark_app` and leaving\nthe rest of the config untouched. A one-field affordance over the raw TOML\neditor; the affected app restarts on broadcast and resolves the new binding.\nA non-empty value must name a registered Lark app, else the app would fail\nits next build — so we reject it up front.",
        "operationId": "set_app_lark_app",
        "parameters": [
          {
            "name": "app",
            "in": "path",
            "description": "Registered app name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LarkAppReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Binding set (or cleared); the app restarts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LarkAppAck"
                }
              }
            }
          },
          "400": {
            "description": "lark_app is not a registered Lark app, or the section is not a table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such registered app",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/lark-apps": {
      "get": {
        "tags": [
          "console"
        ],
        "summary": "`GET /api/lark-apps` — the registry, `app_secret` redacted to a boolean.",
        "operationId": "list_lark_apps",
        "responses": {
          "200": {
            "description": "Registered Lark apps (secrets redacted)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LarkAppsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "post": {
        "tags": [
          "console"
        ],
        "summary": "`POST /api/lark-apps` — live-test the credentials, then write the entry.\nNothing is persisted unless the test passes.",
        "operationId": "upsert_lark_app",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credentials valid and saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpsertAck"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or failed credential test",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/lark-apps/test": {
      "post": {
        "tags": [
          "console"
        ],
        "summary": "`POST /api/lark-apps/test` — dry-run a credential check without saving.\nAlways `200`; the body's `ok` flag carries the verdict.",
        "operationId": "test_lark_app",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credential-test verdict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/lark-apps/{name}": {
      "delete": {
        "tags": [
          "console"
        ],
        "summary": "`DELETE /api/lark-apps/{name}` — drop an entry. Apps still referencing it\nwill fail their next build (surfaced as Errored in the console).",
        "operationId": "delete_lark_app",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Registry entry name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entry removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "404": {
            "description": "No such entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Could not write the config file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppActionSpec": {
        "type": "object",
        "description": "Wire mirror of `larkstack_core::ActionSpec`.",
        "required": [
          "name",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "params": {
            "type": [
              "object",
              "null"
            ],
            "description": "Free-form JSON descriptor the UI renders a form from; omitted when the\naction takes no parameters."
          }
        }
      },
      "AppKind": {
        "type": "string",
        "description": "Wire mirror of `larkstack_core::Kind`.",
        "enum": [
          "integration",
          "automation"
        ]
      },
      "AppManifest": {
        "type": "object",
        "description": "Wire mirror of `larkstack_core::Manifest`, plus the live `enabled` flag read\nfrom config (the manifest itself is static; `enabled` is operator state).",
        "required": [
          "name",
          "kind",
          "description",
          "enabled",
          "actions"
        ],
        "properties": {
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppActionSpec"
            }
          },
          "description": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the supervisor is configured to run this app (`[<name>].enabled`)."
          },
          "kind": {
            "$ref": "#/components/schemas/AppKind"
          },
          "lark_app": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `[lark-apps.<name>]` this app binds for Lark credentials\n(`[<name>].lark_app`), if any. Editable via `PUT /api/config/{app}/lark-app`."
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AppsResponse": {
        "type": "object",
        "description": "Body of `GET /api/apps`.",
        "required": [
          "apps"
        ],
        "properties": {
          "apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppManifest"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Error body shared by every non-2xx JSON response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "LarkAppAck": {
        "type": "object",
        "required": [
          "ok",
          "name"
        ],
        "properties": {
          "lark_app": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
      "LarkAppReq": {
        "type": "object",
        "properties": {
          "lark_app": {
            "type": [
              "string",
              "null"
            ],
            "description": "The `[lark-apps.<name>]` to bind for this app's Lark credentials, or\n`null`/empty to unbind (falling back to env / inline `[<app>.lark]`)."
          }
        }
      },
      "LarkAppView": {
        "type": "object",
        "description": "One registry entry, with the secret redacted to a flag.",
        "required": [
          "name",
          "app_id",
          "base_url",
          "has_secret"
        ],
        "properties": {
          "app_id": {
            "type": "string"
          },
          "base_url": {
            "type": "string"
          },
          "has_secret": {
            "type": "boolean",
            "description": "Whether an `app_secret` is stored (the secret itself is never returned)."
          },
          "name": {
            "type": "string"
          }
        }
      },
      "LarkAppsResponse": {
        "type": "object",
        "description": "Body of `GET /api/lark-apps`.",
        "required": [
          "lark_apps"
        ],
        "properties": {
          "lark_apps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LarkAppView"
            }
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Generic `{ \"ok\": true }` success body.",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "TestReq": {
        "type": "object",
        "required": [
          "app_id",
          "app_secret"
        ],
        "properties": {
          "app_id": {
            "type": "string"
          },
          "app_secret": {
            "type": "string"
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TestResult": {
        "type": "object",
        "description": "Body of `POST /api/lark-apps/test`: `ok` plus either the token lifetime or\nthe failure reason.",
        "required": [
          "ok"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Failure reason, on error."
          },
          "expire": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Token lifetime in seconds, on success.",
            "minimum": 0
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
      "UpsertAck": {
        "type": "object",
        "description": "Body of a successful `POST /api/lark-apps`.",
        "required": [
          "ok",
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
      "UpsertReq": {
        "type": "object",
        "required": [
          "name",
          "app_id",
          "app_secret"
        ],
        "properties": {
          "app_id": {
            "type": "string"
          },
          "app_secret": {
            "type": "string"
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lk_session"
      }
    }
  },
...
}